move and fix job/step names
