+++ service-trigger/src/main/java/com/flores/dev/springbatch/controllers/ServiceJobLauncher.java
@@ -1,4 +1,4 @@
-package com.flores.dev.springbatch;
+package com.flores.dev.springbatch.controllers;
 
 import java.util.Properties;
